fix: revert IconName move
