--- src/components/Badge/Badge.types.ts
@@ -1,1 +1,0 @@
-export type BadgeType = 'default' | 'success' | 'error';
