+++ src/components/CTA/CTA.tsx
@@ -4,9 +4,10 @@ import classnames from 'classnames';
 
 import { BlockHeading } from '../../content-blocks/BlockHeading/BlockHeading';
 import { BlockRichText } from '../../content-blocks/BlockRichText/BlockRichText';
-import { ButtonAction, DefaultProps, HeadingType, IconName } from '../../types';
+import { ButtonAction, DefaultProps, HeadingType } from '../../types';
 import { Button, ButtonProps } from '../Button/Button';
 import { ButtonType } from '../Button/Button.types';
+import { IconName } from '../Icon/Icon.types';
 
 import './CTA.scss';
 
@@ -18,7 +19,7 @@ export interface CTAProps extends DefaultProps {
 	buttonIcon?: IconName;
 	buttonType?: ButtonType;
 	buttonAction: ButtonAction;
-	buildAndNavigate: (buttonAction: ButtonAction) => void;
+	navigate: (buttonAction: ButtonAction) => void;
 }
 
 export const CTA: FunctionComponent<CTAProps> = ({
@@ -30,13 +31,13 @@ export const CTA: FunctionComponent<CTAProps> = ({
 	buttonIcon,
 	buttonType = 'secondary',
 	buttonAction,
-	buildAndNavigate,
+	navigate,
 }) => {
 	const buttonProps: ButtonProps = {
 		label: buttonLabel,
 		title: buttonLabel,
 		ariaLabel: buttonLabel,
-		onClick: () => buildAndNavigate(buttonAction),
+		onClick: () => navigate(buttonAction),
 		icon: buttonIcon,
 		type: buttonType,
 	};
