+++ src/types/index.ts
@@ -22,196 +22,3 @@ export interface ButtonAction {
 	type: ContentPickerType;
 	value: string | number;
 }
-
-export type IconName =
-	| 'activity'
-	| 'add'
-	| 'alert-circle'
-	| 'alert-octagon'
-	| 'alert-triangle'
-	| 'archive'
-	| 'arrow-down-circle'
-	| 'arrow-down-left'
-	| 'arrow-down-right'
-	| 'arrow-down'
-	| 'arrow-left'
-	| 'arrow-right'
-	| 'arrow-up'
-	| 'caret-down'
-	| 'caret-left'
-	| 'caret-right'
-	| 'caret-up'
-	| 'chevron-down'
-	| 'chevron-left'
-	| 'chevron-right'
-	| 'chevron-up'
-	| 'chevrons-down'
-	| 'chevrons-left'
-	| 'chevrons-left-and-right'
-	| 'chevrons-right'
-	| 'chevrons-up'
-	| 'chevrons-up-and-down'
-	| 'bar-chart'
-	| 'bell-off'
-	| 'bell'
-	| 'book-open'
-	| 'book'
-	| 'bookmark-filled'
-	| 'bookmark'
-	| 'box'
-	| 'briefcase'
-	| 'calendar'
-	| 'camera-off'
-	| 'camera'
-	| 'check-circle'
-	| 'check-square'
-	| 'check'
-	| 'circle'
-	| 'clipboard'
-	| 'clock'
-	| 'close'
-	| 'code'
-	| 'collection'
-	| 'contract'
-	| 'copy'
-	| 'credit-card'
-	| 'crop'
-	| 'crosshair'
-	| 'klascement'
-	| 'smartschool'
-	| 'database'
-	| 'delete'
-	| 'download-cloud'
-	| 'download'
-	| 'edit-2'
-	| 'edit-3'
-	| 'edit'
-	| 'external-link'
-	| 'eye-off'
-	| 'eye'
-	| 'fast-forward'
-	| 'file-minus'
-	| 'file-plus'
-	| 'file-text'
-	| 'file'
-	| 'filter'
-	| 'flag'
-	| 'folder-minus'
-	| 'folder-plus'
-	| 'folder'
-	| 'fullscreen'
-	| 'gift'
-	| 'globe'
-	| 'hash'
-	| 'headphone'
-	| 'heart-filled'
-	| 'heart'
-	| 'help-circle'
-	| 'home'
-	| 'image'
-	| 'inbox'
-	| 'info'
-	| 'layout'
-	| 'link-2'
-	| 'link'
-	| 'list'
-	| 'lock'
-	| 'log-in'
-	| 'log-out'
-	| 'mail'
-	| 'maximize-2'
-	| 'maximize'
-	| 'menu'
-	| 'message-circle'
-	| 'message-square'
-	| 'mic'
-	| 'minimize-2'
-	| 'minimize'
-	| 'minus-circle'
-	| 'more-horizontal'
-	| 'move'
-	| 'circle-check'
-	| 'circle-info'
-	| 'circle-warning'
-	| 'music'
-	| 'notification'
-	| 'notifications'
-	| 'octagon'
-	| 'package'
-	| 'paperclip'
-	| 'pause-add'
-	| 'pause-circle'
-	| 'pause'
-	| 'percent'
-	| 'pie-chart'
-	| 'play-circle'
-	| 'play-small'
-	| 'play'
-	| 'plus-circle'
-	| 'plus'
-	| 'power'
-	| 'printer'
-	| 'radio'
-	| 'refresh-ccw'
-	| 'refresh-cw'
-	| 'repeat'
-	| 'rewind'
-	| 'rotate-ccw'
-	| 'rotate-cw'
-	| 'rss'
-	| 'scissors'
-	| 'search'
-	| 'send'
-	| 'settings'
-	| 'share-2'
-	| 'share'
-	| 'shuffle'
-	| 'sidebar'
-	| 'skip-back'
-	| 'skip-forward'
-	| 'slash'
-	| 'sliders'
-	| 'instagram'
-	| 'twitter'
-	| 'square'
-	| 'star'
-	| 'stop-circle'
-	| 'subtitles'
-	| 'tag'
-	| 'target'
-	| 'terminal'
-	| 'thumbs-down'
-	| 'thumbs-up'
-	| 'trash-2'
-	| 'trending-down'
-	| 'trending-up'
-	| 'triangle'
-	| 'tv'
-	| 'type'
-	| 'underline'
-	| 'unlock'
-	| 'upload'
-	| 'user-check'
-	| 'user-plus'
-	| 'user-x'
-	| 'user'
-	| 'users'
-	| 'video-off'
-	| 'video'
-	| 'volume-1'
-	| 'volume-2'
-	| 'volume-mute'
-	| 'volume-x'
-	| 'volume'
-	| 'align-center'
-	| 'align-justify'
-	| 'align-left'
-	| 'align-right'
-	| 'bold'
-	| 'italic'
-	| 'x-circle'
-	| 'x-square'
-	| 'x'
-	| 'youtube'
-	| 'zoom-in'
-	| 'zoom-out';
